% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_effect_size_utils.R
\name{calculate_absolute_es}
\alias{calculate_absolute_es}
\title{Calculate Absolute Effect Size}
\usage{
calculate_absolute_es(data, idv = "id", dv = "y", iv = "condition", params)
}
\arguments{
\item{data}{The dataset of a specific participant, arranged according to the independent variable ('iv')}

\item{idv}{The name of the participant identifier column.}

\item{dv}{The dependent variable to apply the summary function (summary_function) to.  For multiple dependent variables use a string list with the names of each dependent variable (e.g., c('dv1','dv2')),}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable (dv) is expected to differ.}

\item{params}{A list of parameters used by the function to calculate absolute effect size. Includes:
\itemize{
  \item summary_function - The summary function applied to the dependent variable(s), 'dv', under each split of the data.
}}
}
\value{
the function returns the absolute effect size for the given data
}
\description{
The function gets the absolute effect size for a specific participant.
Called from 'test_absolute_es'
}
