% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional_effect_utils.R
\name{calculate_directional_effect}
\alias{calculate_directional_effect}
\title{Calculate Directional Effect}
\usage{
calculate_directional_effect(
  data,
  idv = "id",
  dv = "y",
  iv = "condition",
  params
)
}
\arguments{
\item{data}{The dataset to analyze.}

\item{idv}{The name of the participant identifier column.}

\item{dv}{The dependent variable to apply the summary function (summary_function) to.  For multiple dependent variables use a string list with the names of each dependent variable (e.g., c('dv1','dv2')),}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable (dv) is expected to differ.}

\item{params}{A list of parameters used by the function to calculate the effect. Includes:
\itemize{
  \item summary_function - The summary function applied to the dependent variable(s), 'dv', under each split of the data.
}}
}
\value{
the function returns the mean consistency of signs for the given data
}
\description{
The function calculates the directional effect analysis to the dataset of a specific participant.
Called from 'get_true_score'
}
