% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_utils.R
\name{create_classification_params}
\alias{create_classification_params}
\title{Create Parameters For Sign Consistency}
\usage{
create_classification_params(model = NA, K = NA, handleImbalance = NA)
}
\arguments{
\item{model}{-}
}
\value{
a list of parameters that includes both arguments.
}
\description{
The function creates a list of paramteres to be later passed to the sign consistency function.
}
