% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional_effect_utils.R
\name{create_directional_effect_params}
\alias{create_directional_effect_params}
\title{Create Parameters for Directional Effect}
\usage{
create_directional_effect_params(summary_function, is_null_dist = FALSE)
}
\arguments{
\item{summary_function}{The summary function applied to the dependent variable(s), 'dv' under each split of the data.}

\item{is_null_dist}{Indicating whether params should include an avoid shuffling entry.
This parameter is used when testing for a significant directional effect.
In this test we do not shuffle labels of the independent variable, we simply randomly assign a sign for the effect.}
}
\value{
a list of parameters that includes both arguments.
}
\description{
The function creates a list of parameters to be later passed to the directional effect analysis function.
}
