% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_data.R
\name{create_sample_data}
\alias{create_sample_data}
\title{Title}
\usage{
create_sample_data(
  p_mean,
  p_sd,
  seed = 1,
  N = 30,
  trialsPerCnd = 100,
  wSEsd = 2
)
}
\arguments{
\item{p_mean}{the effect's population mean}

\item{p_sd}{the standard deviation of the population's effect}

\item{seed}{- a seed to use when generating the resulting data frame}

\item{N}{- the number of simulated participants}

\item{trialsPerCnd}{- the number of trials per condition}

\item{wSEsd}{- the standard deviation of within participant (and condition) measures (error term)}
}
\value{
a data frame with three columns: id (participant id), 'condition' (condition label), and 'var' (the dependent variable)
}
\description{
Title
}
