% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_consistency_utils.R
\name{create_sign_consistency_params}
\alias{create_sign_consistency_params}
\title{Create Parameters For Sign Consistency}
\usage{
create_sign_consistency_params(nSplits, max_resampling, summary_function)
}
\arguments{
\item{nSplits}{- The number of random splits to analyze for the estimation of sign consistency.}

\item{max_resampling}{The maximal number consequent invalid summary function return values (see the documentation of the 'summary_function' argument) to use before ignoring the results of a split iteration.}

\item{summary_function}{The summary function applied to the dependent variable(s), 'dv' under each split of the data.}
}
\value{
a list of parameters that includes both arguments.
}
\description{
The function creates a list of parameters to be later passed to the sign consistency function.
}
