% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_directional_tests.R
\name{get_absolute_es}
\alias{get_absolute_es}
\title{Get Absolute Effect Size}
\usage{
get_absolute_es(
  data,
  idv = "id",
  dv = "rt",
  iv = "condition",
  summary_function = lsr::cohensD
)
}
\arguments{
\item{data}{The dataset to analyze.}

\item{idv}{The name of the participant identifier column.}

\item{dv}{The dependent variable to apply the summary function (summary_function) to.  For multiple dependent variables use a string list with the names of each dependent variable (e.g., c('dv1','dv2')),}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable ('dv') is expected to differ.}

\item{summary_function}{The summary function to apply to the dependent variables ('dv') as a function of the independent variable ('iv') for each participant ('idv').}
}
\value{
A list including the results of the function
\itemize{
  \item statistic - The average absolute effect size across all participants.
  \item absolute_es_per_id - Absolute effect size for each participant.
}
}
\description{
The function returns the average absolute effect size across participants.
The function accepts a dataset in long format with specific columns: identifier ('idv'), independent and dependent variables (iv and dv, respectively).
All levels of the independent variable must be included under each identifier.
}
\seealso{
[signcon::test_absolute_es()] which uses this function to test the significance of the group-level absolute effect size.
}
