% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_boot_ci}
\alias{get_boot_ci}
\title{Get Boot CI
The function gets participant-level scores, and calculates a bootstrapped confidence
interval (CI) on the group-level statistic}
\usage{
get_boot_ci(scores, ci_level = 95, ci_reps = 0)
}
\arguments{
\item{scores}{participant-level scores}

\item{ci_level}{- The confidence level (in percents, e.g. setting the argument to 50 generates a 50% CI)
to use when computing the bootstrapped confidence interval on the group-level
statistic (see the return value 'ci', and the argument 'ci_reps').
The default value of this argument is 95, that would lead to computing the 95% confidence interval for
the group-level statistic.}

\item{ci_reps}{- The number repetitions to use when computing the bootstrapped confidence interval
around the group-level statistic.
The default value of this argument is zero, which would lead to not computing the confidence interval at all.}
}
\value{
a vector with the ci values (lower bound, upper bound) according to the given
confidence level ('ci_level'). If 'ci_reps' is set to zero, the function returns NA.
}
\description{
Get Boot CI
The function gets participant-level scores, and calculates a bootstrapped confidence
interval (CI) on the group-level statistic
}
