% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_utils.R
\name{get_classifier_accuracy}
\alias{get_classifier_accuracy}
\title{Get Classifier Accuracy}
\usage{
get_classifier_accuracy(
  data,
  idv = "id",
  dv = "y",
  iv = "condition",
  K,
  handleImbalance
)
}
\arguments{
\item{data}{The dataset of a specific individual, arranged according to the independent variable ('iv')}

\item{idv}{The name of the subject identifier column.}

\item{dv}{The names of the dependent variables columns to classify conditions according to. For multiple dependent variables use a string list with the names of each dependent variable (e.g., c('dv1','dv2')).}

\item{iv}{The name of the independent variable column - the condition to classify,
indicating the different levels under which the dependent variables (dv) are expected to differ.}

\item{K}{- the number of folds to use when calculating the performance of the classifier.
If K is set to 'NA', the function set it to the number of observations of the minority class.}

\item{handleImbalance}{- A Boolean indicating whether to adjust class imbalance (using different weight for each label)}
}
\value{
the function returns the trained classifier accuracy rate
}
\description{
The function calculates the cross-validated classification accuracy for the condition labels, by the dependent variable(s). for a specific participant's data
Called from 'classify_conditions'
}
