% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_condition_classification}
\alias{get_condition_classification}
\title{The function returns the mean classification accuracy of condition labels according to a dependent variable of each participant's data.}
\usage{
get_condition_classification(
  data,
  idv = "id",
  dv = "rt",
  iv = "condition",
  k = NA,
  classifier = NA
)
}
\arguments{
\item{data}{The dataset to analyze}

\item{dv}{The depedent variable to apply the summary function (summary_function) to.}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable (dv) is expected to differ.}

\item{k}{The number of folds to use when splitting the data of each participant to train and test datasets. The default value (NA), will result in using the maximal number of folds possible (k equals the frequency of the less frequent label).}

\item{classifier}{The classifier to train. The default value (NA) will result in creating a SVM classifier with linear kernel.}

\item{id}{An identifier of a specific participant.}
}
\value{
A list including the results of the function
\itemize{
  \item accuracy - The mean classification accuracy across paricipants and folds.
  \item accuracy_per_id - The classification accuracy across folds, for each participant.
  \item prediction_results - A table including a 'k' column, indicating the fold index, a 'test' column, indicating the label to predict, and a 'prediction' column indicating the prediction accuracy for the test.
}
}
\description{
The function accepts a dataset in long format with specific columns: identifier (id), independent and dependent variables (iv and dv, respectively).
For each paricipant, the function calcualtes the k-fold cross-validated accuracy of a classifier trained to predict the labels of levels under the independent variable (iv) based on the dependent variable (dv).
Then, it returns the mean accuracy of individual paricipants classifiers.
All levels of the independent variable must be included under each identifier.
}
\seealso{
[PACKAGE_NAME::test_condition_classification()] which uses this function to test for the statistical significance of sign consistency.
}
