% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_corrected_pval}
\alias{get_corrected_pval}
\title{Get Corrected P Value}
\usage{
get_corrected_pval(statistic, null_dist)
}
\arguments{
\item{statistic}{the statistic obtained from observed scores}

\item{null_dist}{the null distribution to compare the statistic against}
}
\value{
The (one-sided) corrected p-vale
}
\description{
The function applies a p-value correction (see Phipson & Smyth, 2010)
for an empirical p-value obtained in the permutation procedures used in the package.
}
