% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional_tests.R
\name{get_directional_effect}
\alias{get_directional_effect}
\title{Get Directional Effect}
\usage{
get_directional_effect(
  data,
  idv = "id",
  dv = "rt",
  iv = "condition",
  summary_function = NULL
)
}
\arguments{
\item{data}{The dataset to analyze.}

\item{idv}{The name of the participant identifier column.}

\item{dv}{The dependent variable to apply the summary function (summary_function) to.  For multiple dependent variables use a string list with the names of each dependent variable (e.g., c('dv1','dv2')),}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable ('dv') is expected to differ.}

\item{summary_function}{The summary function to apply to the dependent variables ('dv') under each level of the independent variable ('iv') for each participant ('idv').
This function should map a data frame to a number: {data.frame} -> numeric (e.g. function(df) {mean(as.matrix(df[,dv]))}, which is the default summary function).}
}
\value{
A list including the results of the function
\itemize{
  \item statistic - The average effect across all participants.
  \item effect_per_id - An effect score for each participant.
}
}
\description{
The function returns the directional effect (comparing two conditions, under the independent variable, iv) according to the summary function (summary_function).
The function accepts a dataset in long format with specific columns: identifier ('idv'), independent and dependent variables (iv and dv, respectively).
For each participant, the function calculates the difference score between two conditions.
The difference score is calculated by subtracting the result of applying the summary function (summary_function) to the 2nd level if the independent variable, iv, from the 1st level.
All levels of the independent variable must be included under each identifier.
}
\seealso{
[weaknull::test_directional_effect()] which uses this function to test the significance of the group-level effect.
}
