% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_null_distribution}
\alias{get_null_distribution}
\title{Get Null Distribution
The function returns the null distribution according to the 'statistic' and 'configuration' arguments.}
\usage{
get_null_distribution(
  data,
  idv = "id",
  dv = "rt",
  iv = "condition",
  params,
  f,
  perm_repetitions = 25,
  null_dist_samples = 10000,
  preprocessFs = c(),
  preprocessArgs = c()
)
}
\arguments{
\item{data}{the data of a specific individual, arranged according to the independent variable ('iv')}

\item{idv}{The name of the subject identifier column.}

\item{dv}{The dependent variable to apply the summary function (summary_function) to.}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable (dv) is expected to differ.}

\item{preprocessFs}{vector of functions to apply to the data for preprocessing}

\item{preprocessArgs}{vector of arguments for the preprocessing functions}
}
\value{

}
\description{
Get Null Distribution
The function returns the null distribution according to the 'statistic' and 'configuration' arguments.
}
