% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_null_distribution_perm}
\alias{get_null_distribution_perm}
\title{Get Null Distribution Perm}
\usage{
get_null_distribution_perm(
  data,
  idv = "id",
  dv = "rt",
  iv = "condition",
  params,
  f,
  perm_repetitions = 25,
  null_dist_samples = 10000,
  preprocessFs = c(),
  preprocessArgs = c()
)
}
\arguments{
\item{data}{the data of a specific participant, arranged according to the independent variable ('iv')}

\item{idv}{The name of the participant identifier column.}

\item{dv}{The names of the dependent variable(s) to apply the summary function (summary_function) to. For multiple dependent variables use a string list with the names of each dependent variable (e.g., c('dv1','dv2')),}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable (dv) is expected to differ.}

\item{params}{Configuration for the function to apply to the data of each participant ('f')}

\item{f}{The function to apply to the data of each participant, returning the value of interest for the analysis.}

\item{perm_repetitions}{The number of permutation repetitions for each participant}

\item{null_dist_samples}{The number of samples that comprise the output null distribution}

\item{preprocessFs}{An ordered list of functions to apply to the dataset before starting with the analysis.
The default value of the argument (empty list), will result in shuffling the labels of the independent variable, iv.
The function accepts a 'data' argument (a data frame including the data of an individual),
and an 'args' argument which includes all additional arguments needed to run the preprcessing function.}

\item{preprocessArgs}{An ordered list of function arguments, to be used when invoking the 'preprocessFs' (in order, meaning preprocessFs[i](preprocessArgs[i]) will be invoked for each i).
For example, if the preprocessing function accepts arguments 'a','b' as inputs, preprocessArgs should be set to list(a,b)}
}
\value{
A distribution of mean score values computed according to a 'null' effect condition
}
\description{
The function returns the null distribution according to the 'params' argument, and permutation analysis configuration.
}
