% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_scores_per_participant}
\alias{get_scores_per_participant}
\title{Get Scores Per Participant}
\usage{
get_scores_per_participant(
  data,
  idv = "id",
  dv = "rt",
  iv = "condition",
  preprocessFs = c(),
  preprocessArgs = c(),
  params,
  f
)
}
\arguments{
\item{data}{the dataset of a all participants to analyze.}

\item{idv}{The name of the participant identifier column.}

\item{dv}{The names of the dependent variable(s) to apply the summary function (summary_function) to.}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable(s), 'dv', is expected to differ. For multiple dependent variables use a string list with the names of each dependent variable (e.g., c('dv1','dv2')),}

\item{preprocessFs}{An ordered list of functions to apply to the dataset before starting with the analysis.}

\item{preprocessArgs}{An ordered list of function arguments, to be used when invoking the 'preprocessFs' (in order, meaning preprocessFs[i](preprocessArgs[i]) will be invoked for each i).}

\item{params}{Configuration for the function to apply to the data of each participant ('f').}

\item{f}{The function to apply to the data to compute the score of interest for each participant.}
}
\value{
The function returns a data frame, mapping the 'idv' column with a 'score' columns of the scores of each participant.
}
\description{
The function computes and returns the true score of the analysis to conduct for each participant,
as specified by the 'f' and 'params' arguments.
}
