% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_scores_per_subject}
\alias{get_scores_per_subject}
\title{Get True Scores
The function computes and returns the true score of the analysis to conduct for each individual,
as specified by the 'f' and 'params' arguments.}
\usage{
get_scores_per_subject(
  data,
  idv = "id",
  dv = "rt",
  iv = "condition",
  params,
  f,
  preprocessFs,
  preprocessArgs
)
}
\arguments{
\item{data}{the dataset of a all individuals to analyze.}

\item{idv}{The name of the subject identifier column.}

\item{dv}{The dependent variable to apply the summary function (summary_function) to.}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable (dv) is expected to differ.}

\item{params}{Configuration for the function to apply to the data of each individual ('f')}

\item{f}{The function to apply to the data to compute the score of interest for each individual.}

\item{preprocessArgs}{}
}
\value{
The function returns a list of the computed scores of each individual
}
\description{
Get True Scores
The function computes and returns the true score of the analysis to conduct for each individual,
as specified by the 'f' and 'params' arguments.
}
