% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_consistency_utils.R
\name{non_overlapping_halves_split}
\alias{non_overlapping_halves_split}
\title{Non overlapping halves split}
\usage{
non_overlapping_halves_split(trial_selection)
}
\arguments{
\item{trial_selection}{indicator vector of the same length as
the number of trials in the relevant condition, with equal
number of trials. The vector contains indicators for whether
each trial is selected for group 0 (FALSE),
or group 1 (TRUE).}
}
\value{
a list with two items:
\itemize{
  \item grp0 - a vector of indicators for whether each trial
  belongs to to the first split.
  \item grp1 - the complementary indicators vector to 'grp0'.
}
}
\description{
The function returns a split of trials into two
distinct (non-overlapping) halves randomly.
}
