% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{prepare_subject_data}
\alias{prepare_subject_data}
\title{Get Subject Data}
\usage{
prepare_subject_data(
  data,
  idv = "id",
  dv = "rt",
  iv = "condition",
  preprocessFs,
  preprocessArgs
)
}
\arguments{
\item{data}{The data of a specific individual, arranged according to the independent variable ('iv')}

\item{idv}{The name of the subject identifier column.}

\item{dv}{The dependent variable to apply the summary function (summary_function) to.}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable (dv) is expected to differ.}

\item{preprocessFs}{A vector of functions to apply to the data for preprocessing}

\item{preprocessArgs}{A vector of arguments for the preprocessing functions}
}
\value{
processed subject data, prepared to be analyzed
}
\description{
Returns processed data for  a specific subject.
}
