% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampled_data_effect.R
\docType{data}
\name{sampled_data_effect}
\alias{sampled_data_effect}
\title{Sampled Data For True Effect Analysis}
\format{
A \code{tibble}, simulating a within subject experiment with a single independent and dependent variables.
The experiment includes 100 participants, 2 conditions, and 200 trials per condition.
The effect was generated by creating a fixed positive effect for all participants.
The tibble includes3 columns, which are:
\describe{
\item{id}{the id of each participant}
\item{condition}{the independent variable}
\item{var}{the dependent measure}
}
}
\usage{
sampled_data_effect
}
\description{
A data structure with a true positive effect for testing purposes, and running examples
}
\keyword{datasets}
