% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampled_data_strong.R
\docType{data}
\name{sampled_data_strong}
\alias{sampled_data_strong}
\title{Sampled Data For Strong Null Analysis}
\format{
A \code{tibble}, simulating a within subject experiment with a single independent and dependent variables.
The experiment includes 100 participants, 2 conditions, and 200 trials per condition.
Strong null effect was generated by creating a fixed null effect for all participants.
The tibble includes3 columns, which are:
\desctibe{
\item{id}{the id of each participant}
\item{condition}{the independent variable}
\item{var}{the dependent measure}
}
}
\usage{
sampled_data_strong
}
\description{
A data structure with a strong null effect for testing purposes, and running examples
}
\keyword{datasets}
