% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampled_data_weak.R
\docType{data}
\name{sampled_data_weak}
\alias{sampled_data_weak}
\title{Sampled Data For Weak Null Analysis}
\format{
A \code{tibble}, simulating a within subject experiment with a single independent and dependent variables.
The experiment includes 100 participants, 2 conditions, and 200 trials per condition.
Weak null effect was generated by creating individual differences in effects:
Half of the subjects have a positive effect and the other half have a negative effect.
The tibble includes3 columns, which are:
\desctibe{
\item{id}{the id of each participant}
\item{condition}{the independent variable}
\item{var}{the dependent measure}
}
}
\usage{
sampled_data_weak
}
\description{
A data structure with a weak null effect for testing purposes, and running examples
}
\keyword{datasets}
