% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_directional_tests.R
\name{test_absolute_es}
\alias{test_absolute_es}
\title{Tests for Absolute Effect Size}
\usage{
test_absolute_es(
  data,
  idv = "id",
  dv = "rt",
  iv = "condition",
  summary_function = lsr::cohensD,
  perm_repetitions = 100,
  null_dist_samples = 10^4
)
}
\arguments{
\item{data}{The dataset to analyze}

\item{idv}{The name of the participant identifier column.}

\item{dv}{The dependent variable to apply the summary function (summary_function) to.  For multiple dependent variables use a string list with the names of each dependent variable (e.g., c('dv1','dv2')),}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable ('dv') is expected to differ .}

\item{summary_function}{The summary function to apply to the dependent variables ('dv') as a function of the independent variable ('iv') for each participant ('idv').}

\item{null_dist_samples}{The number of samples taken from the null distribution.}
}
\value{
A list including the results of the function
\itemize{
  \item p - The p_value of the absolute effect size against a bootstrapped null distribution.
  the p-value is adjusted according to (B + 1) / (M + 1), see Phipson & Smyth, 2010.
  \item statistic - The group-level statistic describing the average absolute effect size
  \item null_dist - A numerical vector of samples of absolute effect sizes under the null hypothesis (no consistent difference in the dependent variable ('dv') between the levels of the independent variable ('iv')).
  \item absolute_es_per_id - Absolute effect sizefor each participant.
}
}
\description{
The function tests whether the absolute effect size is larger than expected by chance, using bootstrapping and permuting each participants' independent variable labels.
The function accepts a dataset in long format with specific columns: identifier ('idv'), independent and dependent variables (iv and dv, respectively).
For each participant, the function calculates the absolute effect size.
All levels of the independent variable must be included under each identifier.
}
\seealso{
[signcon::get_absolute_es()]
}
