% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional_tests.R
\name{test_directional_effect}
\alias{test_directional_effect}
\title{Tests for a Directional Effect}
\usage{
test_directional_effect(
  data,
  idv = "id",
  dv = "rt",
  iv = "condition",
  summary_function = NULL,
  perm_repetitions = 25,
  null_dist_samples = 10000
)
}
\arguments{
\item{data}{The dataset to analyze.}

\item{idv}{The name of the participant identifier column.}

\item{dv}{The dependent variable to apply the summary function (summary_function) to.  For multiple dependent variables use a string list with the names of each dependent variable (e.g., c('dv1','dv2')),}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable (dv) is expected to differ.}

\item{summary_function}{The summary function to apply to the dependent variables (dv) under each level of the independent variable (iv) for each participant (id).
This function should map a data frame to a number: {data.frame} -> numeric (e.g. function(df) {mean(as.matrix(df[,dv]))}, which is the default summary function).}

\item{perm_repetitions}{The number of label shuffling for each participant.}

\item{null_dist_samples}{The number of samples taken from the null distribution.}
}
\value{
A list including the results of the function
\itemize{
  \item p - The p_value of the estimated effect compared with the distribution of effects under the bootstrapped null distribution.
  \item statistic - The group-level statistic describing the average effect across participants.
  \item null_dist - A numerical vector of samples of effects under the null hypothesis (where the effect of each participant is assigned a random sign).
}
}
\description{
The function tests for a group-level directional of a difference score (comparing two conditions, under the independent variable, iv) according to the summary function (summary_function)for a random split of the data, using bootstrapping and permuting each participants' independent variable labels.
The function accepts a dataset in long format with specific columns: identifier (id), independent and dependent variables (iv and dv, respectively).
For each participant, the function calculates the difference score between two conditions.
The difference score is calculated by subtracting the result of applying the summary function (summary_function) to the 2nd level if the independent variable, iv, from the 1st level.
Then, the average effect across participants is tested against a bootstrapped null distribution in which effects are calculated for each participant with a random sign (see Stelzer, J., Chen, Y., & Turner, R., 2013).
All levels of the independent variable must be included under each identifier.
}
\seealso{
[weaknull::get_directional effect()] returns the directional effect of each participant.
}
