% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_utils.R
\name{train_classifier}
\alias{train_classifier}
\title{Train Classifier}
\usage{
train_classifier(
  formula,
  data,
  idv = "id",
  dv = "y",
  iv = "condition",
  model,
  K,
  handleImbalance
)
}
\arguments{
\item{formula}{the formula that defines the classification task (e.g., iv ~ dv, in the general case).}

\item{data}{The dataset of a specific individual, arranged according to the independent variable ('iv')}

\item{idv}{The name of the subject identifier column.}

\item{dv}{The dependent variable to classify conditions according to.}

\item{iv}{Labels of an independent variable - the condition to classify, indicating the different levels under which the dependent variable (dv) is expected to differ.}

\item{model}{- the name of the classification model to use.}

\item{K}{- the number of folds to use when calculating the performance of the classifier. If K is set to 'NA', the function resets it to the number of observations of the minority class.}

\item{handleImbalance}{- a string indicating whether and which class imbalance adjustment to use.}
}
\value{
the function returns the trained classifier accuracy rate
}
\description{
Train Classifier
}
