% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{validate_data}
\alias{validate_data}
\title{Validate Data}
\usage{
validate_data(data, idv, dv, iv)
}
\arguments{
\item{data}{the data of a specific participant, arranged according to the independent variable ('iv')}

\item{idv}{The name of the participant identifier column.}

\item{dv}{the names of the dependent variable(s) to apply the summary function (summary_function) to. For multiple dependent variables use a string list with the names of each dependent variable (e.g., c('dv1','dv2')),}

\item{iv}{labels of an independent variable, indicating the different levels under which the dependent variable (dv) is expected to differ.}
}
\description{
The function checks if the dataset is valid by performing different checks on it.
}
