% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_consistency_utils.R
\name{calculate_sign_consistency}
\alias{calculate_sign_consistency}
\title{Calculate Sign Consistency}
\usage{
calculate_sign_consistency(
  data,
  idv = "id",
  dv = "y",
  iv = "condition",
  params
)
}
\arguments{
\item{data}{The dataset of a specific participant, arranged according to the independent variable ('iv')}

\item{idv}{The name of the participant identifier column.}

\item{dv}{The dependent variable to apply the summary function (summary_function) to.  For multiple dependent variables use a string list with the names of each dependent variable (e.g., c('dv1','dv2')),}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable (dv) is expected to differ.}

\item{params}{A list of parameters used by the function to calculate sign consistency. Includes:
\itemize{
  \item n_splits - The number of random splits to analyze for the estimation of sign consistency.
  \item summary_function - The summary function applied to the dependent variable(s), 'dv', under each split of the data.
  \item null_dist_f - A function that calculates sign consistency score for the individual under the null hypothesis.
  \item max_invalid_reps - The maximal number repetitions in which invalid consistency was computed before returning NA result.
}}
}
\value{
the function returns the mean consistency of signs for the given data
}
\description{
The function applies the sign consistency analysis to the dataset of a specific participant.
Called from 'get_true_score'
}
