% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_consistency_utils.R
\name{calculate_sign_consistency}
\alias{calculate_sign_consistency}
\title{Calculate Sign Consistency}
\usage{
calculate_sign_consistency(
  data,
  idv = "id",
  dv = "y",
  iv = "condition",
  params
)
}
\arguments{
\item{data}{The dataset of a specific individual, arranged according to the independent variable ('iv')}

\item{idv}{The name of the subject identifier column.}

\item{dv}{The dependent variable to apply the summary function (summary_function) to.}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable (dv) is expected to differ.}

\item{params}{A list of parametrs used by the function to calculate sign consistency. Includes:
\itemize{
  \item nSplits - The number of random splits to analyze for the estimation of sign consistencty.
  \item summary_function - The summary function applied to the dependent variable (dv) under each split of the data.
}}
}
\value{

}
\description{
The function applies the sign consistency analysis to the dataset of a specific individual
Called from 'get_true_score'
}
