% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_utils.R
\name{classify_conditions}
\alias{classify_conditions}
\title{Calculate Sign Consistency}
\usage{
classify_conditions(data, idv = "id", dv = "y", iv = "condition", params)
}
\arguments{
\item{data}{The dataset of a specific individual, arranged according to the independent variable ('iv')}

\item{idv}{The name of the subject identifier column.}

\item{dv}{The dependent variable to apply the summary function (summary_function) to.}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable (dv) is expected to differ.}

\item{params}{A list of parameters used by the function to calculate sign consistency. Includes:
\itemize{
  \item model -
}}
}
\value{
The classification accuracy of the classifier in classifying the 'iv' parameter based on the 'dv' parameter.
}
\description{
The function applies the sign consistency analysis to the dataset of a specific individual
Called from 'get_true_score'
}
