% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_utils.R
\name{classify_conditions}
\alias{classify_conditions}
\title{Classify Conditions}
\usage{
classify_conditions(data, idv = "id", dv = "y", iv = "condition", params)
}
\arguments{
\item{data}{The dataset of a specific individual, arranged according to the independent variable ('iv')}

\item{idv}{The name of the subject identifier column.}

\item{dv}{The names of the dependent variables columns to classify conditions according to. For multiple dependent variables use a string list with the names of each dependent variable (e.g., c('dv1','dv2')).}

\item{iv}{The name of the independent variable column - the condition to classify,
indicating the different levels under which the dependent variables ('dv') are expected to differ.}

\item{params}{A list of parameters used by the function to perform the classification task. Includes:
\itemize{
  \item K - the number of folds to use when calculating the performance of the classifier.
  \item handle_imbalance - A Boolean indicating whether to adjust class imbalance (using different weight for each label)
  \item null_dist_f - A function that calculates classification accuracy score for the individual under the null hypothesis
}}
}
\value{
The classification accuracy of the classifier in classifying the 'iv' parameter based on the 'dv' parameter.
}
\description{
The function fits the data of an individual with a linear SVM classifier and than uses of a specific individual.
Called from 'get_true_score'
}
