% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_utils.R
\name{create_classification_params}
\alias{create_classification_params}
\title{Create Parameters For Classification}
\usage{
create_classification_params(K = NA, handleImbalance = NA)
}
\arguments{
\item{K}{- the number of folds to use when calculating the performance of the classifier. The default value is set to the number of observations of the minority class.}

\item{handleImbalance}{- A Boolean indicating whether to adjust class imbalance (using different weight for each label).}
}
\value{
a list of parameters that includes all arguments after applying default values.
}
\description{
The function creates a list of parameters to be later passed to the classification function.
}
