% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_consistency_utils.R
\name{create_sign_consistency_params}
\alias{create_sign_consistency_params}
\title{Create Parameters For Sign Consistency}
\usage{
create_sign_consistency_params(n_splits, summary_function, max_invalid_reps)
}
\arguments{
\item{n_splits}{- The number of random splits to analyze for the estimation of sign consistency.}

\item{summary_function}{The summary function applied to the dependent variable(s), 'dv' under each split of the data.}

\item{max_invalid_reps}{- The maximal number repetitions in which invalid consistency was computed before returning NA result.}
}
\value{
a list of parameters that includes both arguments.
}
\description{
The function creates a list of parameters to be later passed to the sign consistency function.
}
