% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_shuffled_score}
\alias{get_shuffled_score}
\title{Get Shuffled Score}
\usage{
get_shuffled_score(
  idx,
  data,
  idv,
  dv,
  iv,
  preprocessFs,
  preprocessArgs,
  params,
  f
)
}
\arguments{
\item{idx}{the index of the iteration we are currently running (this function is being called iteratively)}

\item{data}{the data of a specific participant, arranged according to the independent variable ('iv')}

\item{idv}{The name of the participant identifier column.}

\item{dv}{the names of the dependent variable(s) to apply the summary function (summary_function) to. For multiple dependent variables use a string list with the names of each dependent variable (e.g., c('dv1','dv2')),}

\item{iv}{labels of an independent variable, indicating the different levels under which the dependent variable (dv) is expected to differ.}

\item{preprocessFs}{a vector of functions to apply to the data for preprocessing}

\item{preprocessArgs}{a vector of arguments for the preprocessing functions}

\item{params}{configuration for the function to apply to the data of each participant ('f')}

\item{f}{the function to apply to the data of each participant, returning the value of interest for the analysis.}
}
\value{
the function returns the score calculated by applying the function 'f' to the data after suffling the labels of the indepdent variable 'iv'.
}
\description{
The function shuffles the labels the independent variable (according to the value of 'iv'),
and returns result of the function f on the shuffled dataset.
}
