% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_directional_tests.R
\name{get_sign_consistency}
\alias{get_sign_consistency}
\title{Get Sign Consistency}
\usage{
get_sign_consistency(
  data,
  idv = "id",
  dv = "rt",
  iv = "condition",
  repetitions = 1000,
  summary_function = base::mean
)
}
\arguments{
\item{data}{The dataset to analyze.}

\item{idv}{The name of the subject identifier column.}

\item{dv}{The dependent variable to apply the summary function (summary_function) to.}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable (dv) is expected to differ.}

\item{repetitions}{The number of repetitions used to estimate sign consistency probability.}

\item{summary_function}{The summary function applied to the dependent variable (dv) under each level of the independent variable (iv) for each participant (id).}
}
\value{
A list including the results of the function
\itemize{
  \item estimate - The mean sign consistency across all participants.
  \item consistency_per_id - Sign consistency estimate for each participant.
}
}
\description{
The function returns the mean probability of a consistent difference score sign for a random split of each participant's data.
The function accepts a dataset in long format with specific columns: identifier (id), independent and dependent variables (iv and dv, respectively).
For each participant, the function calculates the estimated probability that for a random splitting of the data, the summary function (summary_function) returns two values with the same sign when applied to the two halves.
Then, it returns the mean sign consistency probability across participants.
All levels of the independent variable must be included under each identifier.
}
\seealso{
[weaknull::test_sign_consistency()] which uses this function to test the significance of the group-level sign consistency.
}
