% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{prepare_participant_data}
\alias{prepare_participant_data}
\title{Prepare Participant Data}
\usage{
prepare_participant_data(
  data,
  idv = "id",
  dv = "rt",
  iv = "condition",
  preprocessFs,
  preprocessArgs
)
}
\arguments{
\item{data}{The data of a specific participant, arranged according to the independent variable ('iv')}

\item{idv}{The name of the participant identifier column.}

\item{dv}{The names of the dependent variable(s) to apply the summary function (summary_function) to. For multiple dependent variables use a string list with the names of each dependent variable (e.g., c('dv1','dv2')),}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable(s), 'dv', is expected to differ.}

\item{preprocessFs}{A vector of functions to apply to the data for preprocessing}

\item{preprocessArgs}{A vector of arguments for the preprocessing functions}
}
\value{
processed participant data, ready to be analyzed
}
\description{
Returns preprocessed data for a specific participant.
}
