% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_directional_tests.R
\name{test_condition_classification}
\alias{test_condition_classification}
\title{Tests for Classification Accuracy}
\usage{
test_condition_classification(
  data,
  idv = "id",
  dv = "rt",
  iv = "condition",
  K = NA,
  handleImbalance = NA,
  perm_repetitions = 25,
  null_dist_samples = 10000
)
}
\arguments{
\item{data}{The dataset to analyze}

\item{idv}{The name of the participant identifier column.}

\item{dv}{The dependent variable to apply the summary function (summary_function) to.}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable (dv) is expected to differ.}

\item{K}{- the number of folds to use when calculating the performance of the classifier. If K is set to 'NA', the function resets it to the number of observations of the minority class.}

\item{handleImbalance}{- A Boolean indicating whether to adjust class imbalance (using different weight for each label)}

\item{perm_repetitions}{The number of label shuffling for each participant.}

\item{null_dist_samples}{The number of samples taken from the null distribution.}
}
\value{
A list including the results of the function
\itemize{
  \item p - The p_value of the average classification accuracy across participants, under the bootstrapped null distribution.
  \item statistic - The group-level statistic describing the average classification accuracy across participants.
  \item null_dist - A numerical vector of samples of the classifier accuracies under the bootstrapped null distribution.
}
}
\description{
The function tests for a significant average classification accuracy of condition labels according to a dependent variable across participants (without assuming a directional effect), using bootstrapping and permutating each participants' independent variable labels.
The function accepts a dataset in long format with specific columns: identifier (id), independent and dependent variables (iv and dv, respectively).
For each participant, the function calculates the k-fold cross-validated accuracy of a SVM classifier with a linear kernel trained to predict the labels of levels under the independent variable (iv) based on the dependent variable (dv).
Then, the average classification accuracy across participants is tested against a bootstrapped null distribution in which classification accuracy is calculated for each participant after shuffling its independent variable labels(see Stelzer, J., Chen, Y., & Turner, R. (2013)).
All levels of the independent variable must be included under each identifier.
}
\seealso{
[weaknull::get_condition_classification()] returns the condition classification accuracy of each participant in the dataset.

[e1071::svm()] the SVM classifier used in this function
}
