% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_directional_tests.R
\name{test_sign_consistency}
\alias{test_sign_consistency}
\title{Tests for Sign Consistency
The function tests for a consistent sign of a difference score for a random split of the data, using bootstrapping and permutating each participants' independent variable labels.
The function accepts a dataset in long format with specific columns: identifier (id), independent and dependent variables (iv and dv, respectively).
For each participant, the function calculates the estimated probability that for a random splitting of the data, the summary function (summary_function) returns two values with the same sign when applied to the two halves.
Then, the mean sign consistency across participants is tested against a bootstrapped null distribution in which sign consistency probabilities are calculated for each participant after shuffling its independent variable labels (see Stelzer, J., Chen, Y., & Turner, R. (2013)).
All levels of the independent variable must be included under each identifier.}
\usage{
test_sign_consistency(
  data,
  idv = "id",
  dv = "rt",
  iv = "condition",
  summary_function = base::mean,
  repetitions = 1000,
  perm_repetitions = 25,
  null_dist_samples = 10000
)
}
\arguments{
\item{data}{The dataset to analyze}

\item{idv}{The name of the subject identifier column.}

\item{dv}{The dependent variable to apply the summary function (summary_function) to.}

\item{iv}{Labels of an independent variable, indicating the different levels under which the dependent variable (dv) is expected to differ.}

\item{summary_function}{The summary function applied  to the dependent variable (dv) under each level of the independent variable (iv) for each identifier (id).}

\item{repetitions}{The number of repetitions used to estimate the probability of sign consistency.}

\item{perm_repetitions}{The number of label shuffling for each participant.}

\item{null_dist_samples}{The number of samples taken from the null distribution.}
}
\value{
A list including the results of the function
\itemize{
  \item p - The p_value of the estimated sign consistency under the distribution of sign consistency probabilities under the bootstrappednull distribution.
  \item statistic - The group-level statstic describing the mean sign consistency across participants.
  \item null_dist - A numerical vector of samples of sign consistency under the assumption that there is no consistent difference in the dependent variable (dv) between the levels of the independent variable (iv).
}
}
\description{
Tests for Sign Consistency
The function tests for a consistent sign of a difference score for a random split of the data, using bootstrapping and permutating each participants' independent variable labels.
The function accepts a dataset in long format with specific columns: identifier (id), independent and dependent variables (iv and dv, respectively).
For each participant, the function calculates the estimated probability that for a random splitting of the data, the summary function (summary_function) returns two values with the same sign when applied to the two halves.
Then, the mean sign consistency across participants is tested against a bootstrapped null distribution in which sign consistency probabilities are calculated for each participant after shuffling its independent variable labels (see Stelzer, J., Chen, Y., & Turner, R. (2013)).
All levels of the independent variable must be included under each identifier.
}
\seealso{
[PACKAGE_NAME::get_sign_consistency()] which returns the probability of a consistent sign of a difference score for a random split of the data
}
