% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_utils.R
\name{train_classifier}
\alias{train_classifier}
\title{Title}
\usage{
train_classifier(formula, data, labels, model, K, handleImbalance)
}
\arguments{
\item{handleImbalance}{}
}
\value{

}
\description{
Title
}
